Add go solution: buddy-pairs

+++ buddy-pairs/solution.go
@@ -1,11 +1,13 @@
 package kata
 
-func sumProperDivisors(n uint64) uint64 {
+// sumProperDivisors returns the sum of proper divisors of n (divisors other than n).
+func sumProperDivisors(n int) int {
 	if n < 2 {
 		return 0
 	}
-	sum := uint64(1)
-	for i := uint64(2); i*i <= n; i++ {
+	// 1 always divides n and is a proper divisor for n > 1.
+	sum := 1
+	for i := 2; i*i <= n; i++ {
 		if n%i == 0 {
 			sum += i
 			j := n / i
@@ -17,11 +19,11 @@ func sumProperDivisors(n uint64) uint64 {
 	return sum
 }
 
-func Buddy(start, limit uint64) []uint64 {
+func Buddy(start, limit int) []int {
 	for n := start; n <= limit; n++ {
 		m := sumProperDivisors(n) - 1
 		if m > n && sumProperDivisors(m) == n+1 {
-			return []uint64{n, m}
+			return []int{n, m}
 		}
 	}
 	return nil
